Update README.md
the program makes a new file from 5000 lines of the big data file on the hard drive or flash drive. this "new_file" is then made into a Pandas dataFrame and analyzed. Then the "new_file" is removed on line 147 to make way for the next j iterator to take the next 5000 lines for analysis. line 13 has a command to remove "new_file" incase there is such a file still on the hard drive or flash drive from the last run of the program, which may a have crashed, or stopped prematurely, without erasing the previous "new_file".  If there is no "new_file" on the drive, the program will stop with an error message regarding line 13. So then I just would comment-out line 13 with a "#" .